Add src/app/lib/sort-tweets.ts
