+++ src/app/store.ts
@@ -3,16 +3,9 @@ import { Middleware } from 'redux';
 import { createLogger } from 'redux-logger';
 import { TrashboxElement } from '~/lib/clipboard';
 import { Tweet } from '~/lib/tweet';
+import { TweetSort, tweetSortFunction } from './lib/sort-tweets';
 
 // state
-export type TweetSortKey = 'timestamp' | 'username';
-export type SortOrder = 'asc' | 'desc';
-
-export interface TweetSort {
-  key: TweetSortKey;
-  order: SortOrder;
-}
-
 export interface State {
   tweets: Tweet[];
   selectedTweets: Tweet[];
@@ -93,32 +86,3 @@ export const store = configureStore({
 });
 
 export type Store = typeof store;
-
-// sort tweets
-const tweetSortFunction = ({
-  key,
-  order,
-}: TweetSort): ((lhs: Tweet, rhs: Tweet) => number) => {
-  const sign = order === 'asc' ? +1 : -1;
-  switch (key) {
-    case 'timestamp':
-      return (lhs: Tweet, rhs: Tweet) => sign * (lhs.timestamp - rhs.timestamp);
-    case 'username': {
-      const compareString = new Intl.Collator().compare;
-      return (lhs: Tweet, rhs: Tweet) => {
-        const compareUsername = compareString(
-          lhs.author.username,
-          rhs.author.username,
-        );
-        // if the usenames are equal, the newest first
-        return compareUsername !== 0
-          ? sign * compareUsername
-          : rhs.timestamp - lhs.timestamp;
-      };
-    }
-    default: {
-      const _: never = key;
-      return _;
-    }
-  }
-};
